+++ packages/selected-network-controller/tests/SelectedNetworkMiddleware.test.ts
@@ -1,35 +1,70 @@
-import { Messenger } from '@metamask/base-controller';
 import { JsonRpcEngine } from '@metamask/json-rpc-engine';
+import {
+  Messenger,
+  MOCK_ANY_NAMESPACE,
+  type MessengerActions,
+  type MessengerEvents,
+  type MockAnyNamespace,
+} from '@metamask/messenger';
 import type { JsonRpcResponse } from '@metamask/utils';
 
 import { SelectedNetworkControllerActionTypes } from '../src/SelectedNetworkController';
-import type {
-  AllowedActions,
-  AllowedEvents,
-  SelectedNetworkControllerActions,
-  SelectedNetworkControllerEvents,
-} from '../src/SelectedNetworkController';
+import type { SelectedNetworkControllerMessenger } from '../src/SelectedNetworkController';
 import type { SelectedNetworkMiddlewareJsonRpcRequest } from '../src/SelectedNetworkMiddleware';
 import { createSelectedNetworkMiddleware } from '../src/SelectedNetworkMiddleware';
 
-const buildMessenger = () => {
+type AllSelectedNetworkControllerActions =
+  MessengerActions<SelectedNetworkControllerMessenger>;
+
+type AllSelectedNetworkControllerEvents =
+  MessengerEvents<SelectedNetworkControllerMessenger>;
+
+type RootMessenger = Messenger<
+  MockAnyNamespace,
+  AllSelectedNetworkControllerActions,
+  AllSelectedNetworkControllerEvents
+>;
+
+const controllerName = 'SelectedNetworkController';
+
+/**
+ * Constructs the root messenger.
+ *
+ * @returns A root messenger.
+ */
+function getRootMessenger(): RootMessenger {
+  return new Messenger({
+    namespace: MOCK_ANY_NAMESPACE,
+  });
+}
+
+/**
+ * Constructs the selected network controller messenger.
+ *
+ * @param rootMessenger - A root messenger.
+ * @returns A selected network controller messenger.
+ */
+function getSelectedNetworkControllerMessenger(
+  rootMessenger: RootMessenger,
+): SelectedNetworkControllerMessenger {
   return new Messenger<
-    SelectedNetworkControllerActions | AllowedActions,
-    SelectedNetworkControllerEvents | AllowedEvents
-  >();
-};
+    typeof controllerName,
+    AllSelectedNetworkControllerActions,
+    AllSelectedNetworkControllerEvents,
+    RootMessenger
+  >({
+    namespace: controllerName,
+    parent: rootMessenger,
+  });
+}
 
 const noop = jest.fn();
 
 describe('createSelectedNetworkMiddleware', () => {
   it('throws if not provided an origin', async () => {
-    const messenger = buildMessenger();
+    const rootMessenger = getRootMessenger();
     const middleware = createSelectedNetworkMiddleware(
-      messenger.getRestricted({
-        name: 'SelectedNetworkController',
-        allowedActions: [],
-        allowedEvents: [],
-      }),
+      getSelectedNetworkControllerMessenger(rootMessenger),
     );
     const req: SelectedNetworkMiddlewareJsonRpcRequest = {
       id: '123',
@@ -47,14 +82,9 @@ describe('createSelectedNetworkMiddleware', () => {
   });
 
   it('puts networkClientId on request', async () => {
-    const messenger = buildMessenger();
-    const middleware = createSelectedNetworkMiddleware(
-      messenger.getRestricted({
-        name: 'SelectedNetworkController',
-        allowedActions: [],
-        allowedEvents: [],
-      }),
-    );
+    const rootMessenger = getRootMessenger();
+    const messenger = getSelectedNetworkControllerMessenger(rootMessenger);
+    const middleware = createSelectedNetworkMiddleware(messenger);
 
     const req = {
       origin: 'example.com',
@@ -78,20 +108,13 @@ describe('createSelectedNetworkMiddleware', () => {
 
   it('implements the json-rpc-engine middleware interface appropriately', async () => {
     const engine = new JsonRpcEngine();
-    const messenger = buildMessenger();
+    const rootMessenger = getRootMessenger();
+    const messenger = getSelectedNetworkControllerMessenger(rootMessenger);
     engine.push((req: SelectedNetworkMiddlewareJsonRpcRequest, _, next) => {
       req.origin = 'foobar';
       next();
     });
-    engine.push(
-      createSelectedNetworkMiddleware(
-        messenger.getRestricted({
-          name: 'SelectedNetworkController',
-          allowedActions: [],
-          allowedEvents: [],
-        }),
-      ),
-    );
+    engine.push(createSelectedNetworkMiddleware(messenger));
     const mockNextMiddleware = jest
       .fn()
       .mockImplementation((req, res, _, end) => {